+++ tests/conftest.py
@@ -11,6 +11,8 @@ class DummyCollector(pytest.collect.File):
 def pytest_pycollect_makemodule(path, parent):
     if '_py33' in path.basename and version < (3, 3):
         return DummyCollector(path, parent=parent)
+    if '_py37' in path.basename and version < (3, 7):
+        return DummyCollector(path, parent=parent)
     if '_py3' in path.basename and version < (3, 0):
         return DummyCollector(path, parent=parent)
     if '_py2' in path.basename and version >= (3, 0):